Dev to Staging (#222)
Dev to staging PR

+++ routes/courseRoutes.js
@@ -1,5 +1,12 @@
 const router = require('express').Router();
 const errorCodes = require('../helpers/errorCodes');
+const multer = require('multer');
+const dynamicUpload = multer({
+	storage: multer.memoryStorage(),
+	fileFilter: (req, file, cb) => {
+		cb(null, true);
+	}
+});
 
 // TODO: Update subscriber count to check actual value in DB
 
@@ -16,6 +23,9 @@ const { StudentModel } = require('../models/Students');
 
 // This one is deprecated, but it is used on mobile so we can't delete it yet
 const { OldLectureModel } = require('../models/Lecture');
+
+const { createAndSaveCourse, updateAndSaveCourse } = require('../helpers/courseHelpers');
+const { assert } = require('../helpers/error');
 
 const COMP_TYPES = {
 	LECTURE: 'lecture',
@@ -90,7 +100,8 @@ const topFeedbackOptionForCourses = async (courses) => {
 router.get('/', async (req, res) => {
 	try {
 		// Find all courses in the database and convert to plain objects
-		const courses = await CourseModel.find().lean().populate({ 
+
+		const courses = await CourseModel.find().lean().populate({
 			path: 'creator',
 			populate: {
 				path: 'baseUser',
@@ -117,7 +128,7 @@ router.get('/:id', async (req, res) => {
 		}
 
 		// find a course based on it's id
-		const course = await CourseModel.findById(id).lean().populate({ 
+		const course = await CourseModel.findById(id).lean().populate({
 			path: 'creator',
 			populate: {
 				path: 'baseUser',
@@ -208,7 +219,6 @@ router.get('/:courseId/sections/:sectionId', async (req, res) => {
 		return res.status(500).json({ 'error': errorCodes['E0003'] });
 	}
 });
-
 /**
  * This route is deprecated, but it might be used on mobile so we can't delete it yet
  * instead of the old lecture model, we should use the new one
@@ -374,54 +384,9 @@ router.get('/:section_id/exercises', async (req, res) => {
 	res.send(list);
 });
 
-/*** CREATE COURSE ROUTES ***/
+// /*** CREATE COURSE ROUTES ***/
 
-//Create course route
-router.put('/', async (req, res) => {
-	const { title, category, difficulty, description, creator, status, estimatedHours } = req.body;
-
-	const creatorProfile = await ContentCreatorModel.findOne({ baseUser: creator });
-
-	if (!creatorProfile) {
-		return res.status(400).send({ error: errorCodes['E0004'] }); // If user does not exist, return error
-	}
-
-	const id = creatorProfile._id;
-
-	const course = new CourseModel({
-		title: title,
-		category: category,
-		difficulty: difficulty,
-		description: description,
-		//temporarily commented out as login has not been fully implemented yet
-		//_user: req.user.id,
-		creator: id,
-		published: false,
-		dateCreated: Date.now(),
-		dateUpdated: Date.now(),
-		sections: [],
-		status: status,
-		estimatedHours: estimatedHours,
-		rating: 0,
-	});
-
-	try {
-		const result = await course.save({ new: true })
-			//As id is generated at save, we need to .then() and then save
-			.then(savedCourse => {
-				const generatedId = savedCourse._id;
-				savedCourse.coverImg = generatedId + '_c';
-
-				return savedCourse.save();
-			});
-		return res.status(201).send(result);
-	} catch (err) {
-		return res.status(400).send(err);
-	}
-});
-
-// Update Course
-router.patch('/:id', /*requireLogin,*/ async (req, res) => {
+router.patch('/:id', requireLogin, async (req, res) => {
 	const course = req.body;
 	const { id } = req.params;
 
@@ -435,8 +400,7 @@ router.patch('/:id', /*requireLogin,*/ async (req, res) => {
 			estimatedHours: course.estimatedHours,
 			published: course.published,
 			status: course.status,
-			dateUpdated: Date.now(),
-			coverImg: course.coverImg
+			dateUpdated: Date.now()
 		},
 		function (err) {
 			if (err) {
@@ -446,45 +410,111 @@ router.patch('/:id', /*requireLogin,*/ async (req, res) => {
 	);
 	return res.status(200).send(dbCourse);
 });
-// Update section order of a course 
-router.patch('/:id/sections', async (req, res) => {
-	try {
-		const { id } = req.params;
-		const { sections } = req.body;
 
-		// Validate course ID
-		if (!mongoose.Types.ObjectId.isValid(id)) {
-			return res.status(400).send({ error: errorCodes['E0014'], msg: 'Invalid courseID' + id }); // If id is not valid, return error
-		}
 
-		// Validate section IDs
-		for (const sectionId of sections) {
-			if (!mongoose.Types.ObjectId.isValid(sectionId)) {
-				return res.status(400).send({ error: errorCodes['E0014'], msg: 'invalid sectionID' + sectionId }); // If section id is not valid, return error
+
+router.put('/create/new',
+	requireLogin,
+	dynamicUpload.any(),
+	async (req, res) => {
+		try {
+			const { courseData } = req.body;
+			const parsedCourseData = JSON.parse(courseData);
+			const { courseInfo, sections = [], userId } = parsedCourseData;
+
+			const creatorProfile = await ContentCreatorModel.findOne({ baseUser: userId });
+			assert(creatorProfile, errorCodes.E0013);
+
+			// Centralized file mapping
+			const fileMap = req.files.reduce((acc, file) => {
+				acc[file.fieldname] = file;
+				return acc;
+			}, {});
+
+			// Handle cover image
+			if (fileMap['coverImg']) {
+				courseInfo.coverImg = {
+					...courseInfo.coverImg,
+					file: fileMap['coverImg']
+				};
 			}
+
+			// Handle section videos more dynamically
+			sections.forEach((section, sectionIndex) => {
+				section.components.forEach((component, componentIndex) => {
+					const videoKey = `sections[${sectionIndex}].components[${componentIndex}].video`;
+					if (fileMap[videoKey]) {
+						component.video = {
+							...component.video,
+							file: fileMap[videoKey]
+						};
+					}
+				});
+			});
+			const newCourse = await createAndSaveCourse(courseInfo, sections, creatorProfile);
+			assert(newCourse, errorCodes.E1401);
+
+			res.status(201).send(newCourse);
+
+		} catch (e) {
+			console.error(e);
+			res.status(500).send(e.message);
 		}
-
-		// Find the course
-		const course = await CourseModel.findById(id);
-
-		// Check if course exists
-		if (!course) {
-			return res.status(404).json({ error: errorCodes['E0006'] }); // If course not found, return error
-		}
-
-		// Update the sections order
-		course.sections = sections;
-
-		// Save the updated course
-		await course.save();
-
-		// Send response
-		return res.status(200).send(course);
-
-	} catch (error) {
-		return res.status(500).json({ error: errorCodes['E0003'] });
 	}
-});
+);
+
+router.post('/update/:id',
+	requireLogin,
+	dynamicUpload.any(),
+	async (req, res) => {
+		const { id } = req.params;
+
+		try {
+			const { courseData } = req.body;
+			const parsedUpdatedCourse = JSON.parse(courseData);
+			const { courseInfo, sections = [] } = parsedUpdatedCourse;
+
+			const baseCourse = await CourseModel.findOne({ _id: id });
+			assert(baseCourse, errorCodes.E1402);
+
+			// Centralized file mapping
+			const fileMap = req.files.reduce((acc, file) => {
+				acc[file.fieldname] = file;
+				return acc;
+			}, {});
+
+			// Handle cover image
+			if (fileMap['coverImg']) {
+				courseInfo.coverImg = {
+					...courseInfo.coverImg,
+					file: fileMap['coverImg']
+				};
+			}
+
+			// Handle section videos more dynamically
+			sections.forEach((section, sectionIndex) => {
+				section.components.forEach((component, componentIndex) => {
+					const videoKey = `sections[${sectionIndex}].components[${componentIndex}].video`;
+					if (fileMap[videoKey]) {
+						component.video = {
+							...component.video,
+							file: fileMap[videoKey]
+						};
+					}
+				});
+			});
+			const updatedCourseModel = await updateAndSaveCourse(courseInfo, sections, baseCourse);
+			assert(updatedCourseModel, errorCodes.E1412);
+
+			res.status(201).send(updatedCourseModel);
+
+		} catch (e) {
+			console.error(e.message);
+			res.status(500).send(e.message);
+		}
+	}
+);
+
 
 /**
  * Delete course by id
@@ -502,21 +532,24 @@ router.delete('/:id'/*, requireLogin*/, async (req, res) => {
 
 
 	// Get the section array from the course object
-	const sectionIds = course.sections;
+	if (course.sections) {
 
-	// Loop through all sections in course
-	sectionIds.map(async (section_id) => {
+		const sectionIds = course.sections;
 
-		// Get the section object from the id in sectionIds array
-		let section = await SectionModel.findById(section_id);
+		// Loop through all sections in course
+		sectionIds.map(async (section_id) => {
 
-		// Delete all lectures and excercises in the section
-		await LectureModel.deleteMany({ parentSection: section._id });
-		await ExerciseModel.deleteMany({ parentSection: section._id });
+			// Get the section object from the id in sectionIds array
+			let section = await SectionModel.findById(section_id);
 
-		// Delete the section
-		await SectionModel.findByIdAndDelete(section_id);
-	});
+			// Delete all lectures and excercises in the section
+			await LectureModel.deleteMany({ parentSection: section._id });
+			await ExerciseModel.deleteMany({ parentSection: section._id });
+
+			// Delete the section
+			await SectionModel.findByIdAndDelete(section_id);
+		});
+	}
 
 	// Delete the course
 	await CourseModel.findByIdAndDelete(id).catch((err) => res.status(204).send(err));
@@ -526,27 +559,6 @@ router.delete('/:id'/*, requireLogin*/, async (req, res) => {
 	return res.status(200).send('Course Deleted');
 
 });
-
-
-// Update course published status 
-// Status is enum: "published", "draft", "hidden"
-router.patch('/:id/updateStatus', async (req, res) => {
-	const { status } = req.body;
-	const { id } = req.params;
-
-	// find object in database and update title to new value
-	(
-		await CourseModel.findOneAndUpdate(
-			{ _id: id },
-			{ status: status }
-		)
-	).save;
-	const course = await CourseModel.findById(id);
-
-	// Send response
-	res.send(course);
-});
-
 
 
 module.exports = router;
